Create PUSH-POP VS SHIFT-UNSHIFT
In the given code. the push function added the item to the last position of array and pop removes the last indexed item from given array.
on the other hand unshift function added the item or value to the first index position in Array and similarly shift function removes the first indexed positioned value/item.